requirements.txt
crewai==0.12.0
crewai-tools
python-dotenv
